+++ pyGeno/Protein.py
@@ -38,9 +38,10 @@ class Protein(pyGenoRabaObjectWrapper) :
 	def _makeLoadQuery(self, objectType, *args, **coolArgs) :
 		if issubclass(objectType, SNP_INDEL) :
 			f = RabaQuery(objectType, namespace = self._wrapped_class._raba_namespace)
+			coolArgs['species'] = self.genome.species
 			coolArgs['chromosomeNumber'] = self.chromosome.number
-			coolArgs['start'] = self.transcript.start
-			coolArgs['end'] = self.transcript.end
+			coolArgs['start >='] = self.transcript.start
+			coolArgs['start <'] = self.transcript.end
 		
 			if len(args) > 0 and type(args[0]) is types.ListType :
 				for a in args[0] :